fix int (#107)

+++ gluoncv/model_zoo/cifarresnext.py
@@ -47,7 +47,7 @@ class CIFARBlock(HybridBlock):
     def __init__(self, channels, cardinality, bottleneck_width,
                  stride, downsample=False, **kwargs):
         super(CIFARBlock, self).__init__(**kwargs)
-        D = math.floor(channels * (bottleneck_width / 64))
+        D = int(math.floor(channels * (bottleneck_width / 64)))
         group_width = cardinality * D
 
         self.body = nn.HybridSequential(prefix='')
